requirements.txt
absl-py==2.3.1
albumentations==1.3.0
astunparse==1.6.3
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
decord==0.6.0
deepface==0.0.93
dlib==20.0.0
efficientnet_pytorch==0.7.0
einops==0.8.1
face-recognition==1.3.0
face_recognition_models==0.3.0
facenet-pytorch==2.6.0
filelock==3.13.1
fire==0.7.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.58.4
fsspec==2024.6.1
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grad-cam==1.5.5
grpcio==1.73.1
gunicorn==23.0.0
h5py==3.14.0
idna==3.10
imageio==2.37.0
imgaug==0.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.5.1
keras==3.10.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
lz4==4.4.4
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
mtcnn==1.0.0
munch==4.0.0
namex==0.1.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.0
pillow==10.2.0
pretrainedmodels==0.7.4
progress==1.6.1
protobuf==5.29.5
Pygments==2.19.2
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-warmup==0.2.0
pytz==2025.2
PyYAML==6.0.2
qudida==0.0.4
requests==2.32.4
retina-face==0.0.17
rich==14.0.0
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.16.0
setuptools==70.2.0
shapely==2.1.1
six==1.17.0
soupsieve==2.7
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.6.11
timm==0.6.5
torch==2.2.2
torchaudio==2.5.1+cu121
torchvision==0.17.2
tqdm==4.67.1
triton==3.1.0
ttach==0.0.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
vit-pytorch==1.10.1
warmup_scheduler==0.3
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
